+++ src/Services/CRM/Common/Result/AbstractCrmItem.php
@@ -50,6 +50,15 @@ class AbstractCrmItem extends AbstractItem
             case 'ASSIGNED_BY_ID':
             case 'CREATED_BY_ID':
             case 'MODIFY_BY_ID':
+            case 'createdBy':
+            case 'updatedBy':
+            case 'movedBy':
+            case 'begindate':
+            case 'closedate':
+            case 'opportunity':
+            case 'opportunityAccount':
+            case 'taxValueAccount':
+            case 'taxValue':
                 // deal
             case 'LEAD_ID':
             case 'CONTACT_ID':
@@ -58,12 +67,20 @@ class AbstractCrmItem extends AbstractItem
             case 'OWNER_ID':
                 // DealCategoryItem
             case 'SORT':
+            case 'id':
+            case 'categoryId':
+            case 'webformId':
+            case 'assignedById':
+            case 'contactId':
+            case 'lastActivityBy':
                 if ($this->data[$offset] !== '' && $this->data[$offset] !== null) {
                     return (int)$this->data[$offset];
                 }
 
                 return null;
             case 'COMPANY_ID':
+            case 'companyId':
+            case 'mycompanyId':
                 if ($this->data[$offset] !== '' && $this->data[$offset] !== null && $this->data[$offset] !== '0') {
                     return (int)$this->data[$offset];
                 }
@@ -74,7 +91,9 @@ class AbstractCrmItem extends AbstractItem
             case 'HAS_EMAIL':
             case 'HAS_IMOL':
             case 'OPENED':
+            case 'opened':
             case 'IS_MANUAL_OPPORTUNITY':
+            case 'isManualOpportunity':
             case 'CLOSED':
             case 'IS_NEW':
             case 'IS_LOCKED':
@@ -88,6 +107,10 @@ class AbstractCrmItem extends AbstractItem
             case 'BIRTHDATE':
             case 'BEGINDATE':
             case 'CLOSEDATE':
+            case 'createdTime':
+            case 'updatedTime':
+            case 'movedTime':
+            case 'lastActivityTime':
                 if ($this->data[$offset] !== '') {
                     return DateTimeImmutable::createFromFormat(DATE_ATOM, $this->data[$offset]);
                 }
@@ -113,6 +136,9 @@ class AbstractCrmItem extends AbstractItem
                     $items[] = new Phone($phone);
                 }
                 return $items;
+            case 'currencyId':
+            case 'accountCurrencyId':
+                return new Currency($this->data[$offset]);
             default:
                 return $this->data[$offset] ?? null;
         }
@@ -128,7 +154,9 @@ class AbstractCrmItem extends AbstractItem
      */
     protected function getKeyWithUserfieldByFieldName(string $fieldName)
     {
-        $fieldName = self::CRM_USERFIELD_PREFIX . $fieldName;
+        if(!str_starts_with($fieldName, self::CRM_USERFIELD_PREFIX)) {
+            $fieldName = self::CRM_USERFIELD_PREFIX . $fieldName;
+        }
         if (!$this->isKeyExists($fieldName)) {
             throw new UserfieldNotFoundException(sprintf('crm userfield not found by field name %s', $fieldName));
         }